Yippie the compiler is now done with the basic opcodes I have at this moment. After this I will make changes to the documentation because it does not reflect what I am doing right now and after that I will update the python 'compiler' because it is ugly and also misses some stuff

+++ virtual_machine/main.cpp
@@ -53,7 +53,7 @@ int main(int argc, char* argv[]) {
 
     std::vector<uint8_t> flashMemory;
 
-    std::array<uint8_t, 0xFF> registers;
+    std::array<uint16_t, 0xFF> registers;
     registers.fill(0);
 
     for (size_t i = 0; i < fileSize; ++i) {
